Refactor wordclock and snakegame
